Address Thermos review findings
